fix: harden 18 latent panic/overflow sites (audit pass-2 latent findings)
The second audit's latent findings — public functions that panic or overflow on
a valid-typed but extreme input. Not currently reachable from untrusted input,
but defensive hardening toward totality. No behaviour change for valid inputs.

Integer overflow → checked / widened / closed-form:
- math is_prime / factorize: `i*i <= n` → `i <= n/i` (overflow-safe).
- math cassini: return the closed form (-1)^n (the products overflow i64, and
  fib overflows u64, for large n).
- calculator isqrt: checked_mul for the (s+1)² correction.
- calculator lcm: saturating_mul.
- music Interval::compose / Note::transpose / pitch-class naming: widen or
  saturate the adds.
- chess PGN target/source square: validate the file/rank bytes before the u8
  subtraction (a malformed move like "12" would underflow).

Slice / index OOB → bounds / char-boundary guards:
- judicial relation_extractor: `body[abs..]` used a lowercased-body offset on
  the original body (a non-char-boundary if case-folding changed byte lengths)
  and a raw 200-byte cut — guard with .get() and a char-boundary-safe scan.
- xsd temporal: is_char_boundary guard before `&s[len-6..]`.
- probability prob / event_prob: .get() (out-of-range outcome → 0).

Unbounded recursion / allocation → bounds:
- calculator Expr::eval and xml_grammar rhs_parser: depth bound (256), mirroring
  the interpreter / parser.
- math goldbach: test primes directly via is_prime (no O(n) sieve allocation);
  math sieve: guard the n+1 overflow.

(sensor_fusion calibration's assert_eq on raw vs bias length is kept — an
internal precondition, not untrusted input.)

+++ crates/domains/src/natural/music/note.rs
@@ -27,7 +27,8 @@ impl Note {
 
     /// Transpose by semitones. Returns None if out of MIDI range.
     pub fn transpose(&self, semitones: i16) -> Option<Note> {
-        let new = self.0 as i16 + semitones;
+        // Widen to i32 so the add cannot overflow before the range test.
+        let new = self.0 as i32 + semitones as i32;
         if (0..=127).contains(&new) {
             Some(Note(new as u8))
         } else {
